Move editor page to index
